bug(server) install the right pg version
