+++ .gitignore
@@ -1,2 +1,4 @@
 .vscode
 node_modules
+dist
+
