+++ server/controllers/book.js
@@ -40,7 +40,7 @@ export default {
                     }
                 });
             })
-            .catch(error => res.status(400).send(error));
+            .catch(error => res.status(400).send(error, "err"));
     },
 
     // lists books by category
